add updated package.json
