+++ app.js
@@ -1,10 +1,6 @@
 import express from 'express';
 import bodyParser from 'body-parser';
 import userRouter from './server/routes/users';
-import messageRouter from './server/routes/message';
-import specificRouter from './server/routes/specificEmail';
-import deleteRouter from './server/routes/deleteMessage';
-import groupsRouter from './server/routes/groups';
 import swagger from 'swagger-ui-express';
 import yamljs from 'yamljs';
 
@@ -17,11 +13,8 @@ app.use('/doc',swagger.serve, swagger.setup(swaggerDocumment));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 
-app.use('/api/v2/messages',specificRouter);
 app.use('/api/v2/auth/',userRouter);
-app.use('/api/v2/',messageRouter);
-app.use('/api/v2/messages',deleteRouter);
-app.use('/api/V2/',groupsRouter);
+
 
 app.get('/',(req,res)=>{
